Added consequences for deleting an entity
Ingredients can not be deleted unless it is used in no recipes.

Recipes can ALWAYS be deleted.

Chefs can always be deleted. If a chef had recipes, these will be given to a custom default chef. Ideally we could give it a special id like -1 or 0.

I have done some basic teting, and this implementation seems to work properly. However, additional testing would be ideal.

Also, I would like to take a closer look to when exactly each change is saved to the database, as well as which side of each relationship is the owner and such things.

+++ RecipeWebPage/src/main/java/distribuidos/recetas/RecipeWebPage/service/IngredientService.java
@@ -2,6 +2,7 @@ package distribuidos.recetas.RecipeWebPage.service;
 
 import distribuidos.recetas.RecipeWebPage.DTO.IngredientDTO;
 import distribuidos.recetas.RecipeWebPage.entities.Ingredient;
+import distribuidos.recetas.RecipeWebPage.entities.Recipe;
 import distribuidos.recetas.RecipeWebPage.repository.IngredientRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -59,6 +60,11 @@ public class IngredientService {
         Optional<Ingredient> ingredientOptional = ingredientRepository.findById(id);
         if (ingredientOptional.isPresent()) {
             Ingredient ingredient = ingredientOptional.get();
+
+            if (!ingredient.getBestRecipes().isEmpty()) {
+                return null; //only allow deletion if it has no recipes associated
+            }
+
             ingredientRepository.delete(id);
             ingredientRepository.delete(ingredient);
             return ingredient;
